+++ CHANGELOG.md
@@ -13,11 +13,19 @@
 - [ibc-relayer]
   - Enable TLS support for gRPC client ([#877])
 
+### IMPROVEMENTS
+
+- [ibc-relayer]
+  - Update the on-chain IBC client with supporting headers when light client verification
+    does bisection when verifying a header for a client update or a misbehaviour detection ([#673])
+
 ### BREAKING CHANGES
 
 - [ibc-relayer-cli]
   - Removed `--coin-type` option from `keys restore` command. Use `--hd-path` instead. ([#1049])
 
+[#673]: https://github.com/informalsystems/ibc-rs/issues/673
+[#868]: https://github.com/informalsystems/ibc-rs/issues/1049
 [#877]: https://github.com/informalsystems/ibc-rs/issues/877
 [#1049]: https://github.com/informalsystems/ibc-rs/issues/1049